+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Remove Travis CI for now. 
+- fixed `tensor.OfSlice()` throw error due to "Unsupported Go type" (e.g. []float32)
 - added `nn.Path.Paths()` method
 - added `nn.VarStore.Summary()` method
 - fixed incorrect tensor method `ts.Meshgrid` -> `Meshgrid`